requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
databricks-sdk==0.96.0
google-auth==2.49.0
idna==3.11
numpy==2.4.2
pandas==3.0.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
urllib3==2.6.3
yml==0.0.1
